refactor code into sub-dir on main git repo
