+++ zscaler_api_talkers/zia_talker/zia_portaltalker.py
@@ -1,4 +1,4 @@
-from zscaler_helpers.http_calls import HttpCalls
+from zscaler_helpers import HttpCalls
 import time
 from getpass import getpass
