+++ zscaler_api_talkers/zpa_talker/zpa_portaltalker.py
@@ -1,8 +1,6 @@
-import pdb
-
 import requests
 
-from zscaler_helpers.http_calls import HttpCalls
+from zscaler_helpers import HttpCalls
 
 
 class ZpaPortalTalker(object):
