+++ zscaler_api_talkers/zpa_talker/zpa_talker.py
@@ -1,4 +1,4 @@
-from zscaler_helpers.http_calls import HttpCalls
+from zscaler_helpers import HttpCalls
 
 
 class ZpaTalker(object):
