--- zscaler_helpers/__init__.py
@@ -1,10 +1,0 @@
-from logger import setup_logger
-from utilities import _request, get_user_agent
-from http_calls import HttpCalls
-
-__all__ = [
-    'setup_logger',
-    '_request',
-    'get_user_agent',
-    'HttpCalls',
-]
